Move tasks from tasks.py to workflows.py
